ai/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
numpy==1.22.3
opencv-python==4.5.5.64
pymongo==4.1.0
pytz==2022.1
sqlparse==0.2.4
